Saving before experimenting
I'm at a place where I'm fine being at but would like to change how the
page responds to the browser growing and shrinking. I tried a few times,
but each one ruined the page, so this is simply to ensure I have a safe
backup to load up to. This might be a good time to try creating a
seperate branch and experimenting without possibly ruining this one.

The main idea is: when the browser is shrunk to as thin as possible,
there should be a slight margin on the content (especially the word-def)
so that the buttons don't overlap what the viewer (me) is trying to read.
I might try having one container (wrap) around another--this 'another'
holds the content--with margins on the side. But then I want it so that
this container isn't any larger (should be slightly smaller) than the
title; maybe max-width? With one container (and my knowledge so far), I
don't think it's possible, but two might work.