Js reads innerHTML
I decided to implement a feature: after all the words / definitions,
there's a section where (soon) you pick how many words you want to be
randomly selected. Of course, the first step was to get javascript to
read innerHTML (not the tags, but what's inside, like mama, etc). It'll
read up to the very last word, and then stop. The goal is to store (push)
them into an array so that when you hit shuffle, the randomizer will
select random numbers, which will correspond to words, which it will then
display.